move axios call from projects to homepage component and add loader animation
